feat: Refactor test routes and enhance test management
- Updated the test route structure by removing the old tests.js file and integrating its functionality into tests/index.js.
- Changed the endpoint for starting initial tests to improve the test creation logic with MongoDB integration.
- Added user ID tracking and validation for test submissions and retrieval.
- Enhanced the TestSchema to include user answers, completion status, and creation date for better test management.
