+++ server.js
@@ -15,7 +15,7 @@ app.use('/api/progress-analyzer', authenticate, require('./routes/agents/progres
 // Добавляем маршрут для начальной диагностики
 app.use('/api/initial-diagnostics', require('./routes/agents/initial-diagnostics'));
 // Добавляем маршрут для работы с тестами
-app.use('/api/tests', authenticate, require('./routes/agents/tests'));
+app.use('/api/tests', authenticate, require('./routes/agents/tests/index'));
 
 const SECRET = process.env.JWT_SECRET || 'ваш_резервный_секрет';
 const MONGODB_URI = process.env.MONGODB_URI;
